requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
arxiv==2.1.3
astrapy==1.4.1
asttokens==2.4.1
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.7.0
blis==0.7.11
boto3==1.35.5
botocore==1.35.5
bs4==0.0.2
cachetools==5.5.0
cassandra-driver==3.29.1
cassio==0.1.8
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colbert-ai==0.2.19
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.19.2
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
emoji==2.12.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
faiss-cpu==1.8.0.post1
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
Flask==3.0.1
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.10
langchain-astradb==0.3.3
langchain-community==0.2.9
langchain-core==0.2.22
langchain-google-genai==1.0.6
langchain-google-vertexai==1.0.5
langchain-nvidia-ai-endpoints==0.1.1
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.101
language_data==1.2.0
llama-cloud==0.0.15
llama-index==0.10.57
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.57
llama-index-embeddings-azure-openai==0.1.7
llama-index-embeddings-bedrock==0.1.4
llama-index-embeddings-gemini==0.1.7
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post3
llama-index-llms-anthropic==0.1.17
llama-index-llms-azure-openai==0.1.6
llama-index-llms-bedrock==0.1.7
llama-index-llms-gemini==0.1.10
llama-index-llms-openai==0.1.31
llama-index-llms-vertex==0.1.5
llama-index-multi-modal-llms-gemini==0.1.7
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-tools-cassandra==0.1.1
llama-index-vector-stores-astra-db==0.1.7
llama-index-vector-stores-cassandra==0.1.3
llama-parse==0.4.1
lxml==4.9.4
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.5.2
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
openai==1.42.0
ordered-set==4.1.0
orjson==3.10.7
packaging==24.0
pandas==2.2.0
parso==0.8.4
peewee==3.17.6
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pubmedpy==0.0.0
pure_eval==0.2.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.2.0
ragstack==0.0.1
ragstack-ai==1.1.0
ragstack-ai-colbert==1.0.6
ragstack-ai-langchain==1.4.0
ragstack-ai-llamaindex==1.0.8
rapidfuzz==3.9.6
referencing==0.35.1
regex==2024.5.10
requests==2.32.3
requests-futures==1.0.1
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scipy==1.14.1
setuptools==73.0.1
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
streamlit==1.37.1
striprtf==0.0.26
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
TextLoader==0.0.9
thinc==8.2.5
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ucimlrepo==0.0.3
ujson==5.10.0
unstructured==0.14.5
unstructured-client==0.25.5
uritemplate==4.1.1
urllib3==2.2.1
uuid6==2024.1.12
virtualenv==20.26.3
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.5.0
yahooquery==2.3.7
yarl==1.9.4
yfinance==0.2.43
